ENH: eliminated build errors on Win64 by redeclaration as size_t.
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@4489 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/biff/biffbiff.c
@@ -265,7 +265,7 @@ biffGet(const char *key) {
   msg = _bmsgFind(key);
   if (!msg) {
     static const char err[]="[%s] No information for this key!";
-    unsigned int errlen;
+    size_t errlen;
     fprintf(stderr, "%s: WARNING: no information for key \"%s\"\n", me, key);
     errlen = strlen(err)+strlen(key)+1;
     ret = AIR_CALLOC(errlen, char);
